requirements.txt
Flask==0.9
PyBrowserID==0.9.0